Se culmina esta tarea implementando el routing hacia las vistad de creación para cada documento, se integra el dashboard de proyecto y un componente con la versión preliminar del módulo de creación de texto enriquecido. También se creó el servicio a nivel de backend para esta funcionalidad.

Issues: #77 #69
