+++ frontend/src/components/documento/vista-modulos/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+
+export default function VistaModulos(props){
+    const {idProyecto} = props;
+
+    return (
+        <div>
+            <h1>Documento de modulos para el proyecto id={idProyecto}</h1>
+        </div>
+    );
+}
