+++ frontend/src/components/vista-proyecto/DocDiseño.js
@@ -3,19 +3,17 @@ import { makeStyles } from '@material-ui/core/styles';
 import clsx from 'clsx';
 import Card from '@material-ui/core/Card';
 import CardHeader from '@material-ui/core/CardHeader';
-import CardMedia from '@material-ui/core/CardMedia';
 import CardContent from '@material-ui/core/CardContent';
 import CardActions from '@material-ui/core/CardActions';
 import Collapse from '@material-ui/core/Collapse';
 import Avatar from '@material-ui/core/Avatar';
 import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
-import { red } from '@material-ui/core/colors';
-import VisibilityIcon from '@material-ui/icons/Visibility';
-import EditIcon from '@material-ui/icons/Edit';
-import GetAppIcon from '@material-ui/icons/GetApp';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import Grid from '@material-ui/core/Grid';
+import VisibilityIcon from "@material-ui/icons/Visibility";
+import EditIcon from "@material-ui/icons/Edit";
+import GetAppIcon from "@material-ui/icons/GetApp";
+import {Link} from "react-router-dom";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -23,7 +21,7 @@ const useStyles = makeStyles((theme) => ({
     },
     media: {
         height: 0,
-        paddingTop: '100%', // 16:9
+        paddingTop: '50%', // 16:9
     },
     expand: {
         transform: 'rotate(0deg)',
@@ -35,43 +33,44 @@ const useStyles = makeStyles((theme) => ({
     expandOpen: {
         transform: 'rotate(180deg)',
     },
-    avatarR: {
-        backgroundColor: red[500],
+    avatarD: {
+        backgroundColor: '#2DA64E',
     },
-
 }));
 
-export default function RecipeReviewCard() {
+export default function RecipeReviewCard(props) {
+    const {idProyecto} = props;
     const classes = useStyles();
-    const [expandedR, setExpandedR] = React.useState(false);
-    const handleExpandClickR = () => {
-        setExpandedR(!expandedR);
+    const [expandedD, setExpandedD] = React.useState(false);
+
+    const handleExpandClickD = () => {
+        setExpandedD(!expandedD);
     };
 
     return (
         <div>
-                {/* Requisitos */}
+                {/* Diseño */}
                 <Card className={classes.root}>
                     <CardHeader
                         avatar={
-                            <Avatar aria-label="recipe" className={classes.avatarR}>
-                                R
+                            <Avatar aria-label="recipe" className={classes.avatarD}>
+                                D
                             </Avatar>
                         }
-                        title="Documento de Requisitos de Software"
-                        subheader="Última modificación: 18 Julio 2020 "
+                        title="Documento de Diseño de Software"
+                        subheader="última modificación: 18 Julio 2020 "
                     />
                     <CardContent>
                         <Typography paragraph>Avance:</Typography>
                         <Typography variant="body2" color="textSecondary" component="p">
-                            30%
+                            50%
                         </Typography>
                     </CardContent>
                     <CardActions disableSpacing>
                         <IconButton aria-label="ver">
                             <VisibilityIcon/>
                         </IconButton>
-                        <IconButton aria-label="editar">
+                        <IconButton component={Link} to={"/proyecto/"+idProyecto+"/dis"} aria-label="editar">
                             <EditIcon />
                         </IconButton>
                         <IconButton aria-label="descargar">
@@ -79,33 +78,26 @@ export default function RecipeReviewCard() {
                         </IconButton>
                         <IconButton
                             className={clsx(classes.expand, {
-                                [classes.expandOpen]: expandedR,
+                                [classes.expandOpen]: expandedD,
                             })}
-                            onClick={handleExpandClickR}
-                            aria-expanded={expandedR}
+                            onClick={handleExpandClickD}
+                            aria-expanded={expandedD}
                         >
                             <ExpandMoreIcon />
                         </IconButton>
                     </CardActions>
-                    <Collapse in={expandedR} timeout="auto" unmountOnExit>
+                    <Collapse in={expandedD} timeout="auto" unmountOnExit>
                         <CardContent>
                             <Typography paragraph>Descripción:</Typography>
                             <Typography variant="body2" color="textSecondary" component="p">
-                                El documento de requerimientos de software, es el lugar donde se da
-                                descripción a las características y requisitos de un software, producto, programa
-                                o conjunto de programas. Los requisitos se expresan en lenguaje natural, sin
-                                consideraciones ni términos técnicos.
-                                La especificación de requisitos de software es el resultado del levantamiento
-                                de información con el usuario o cliente del producto.
+                                El documento de diseño de software, es una descripción escrita del un producto de software,
+                                en donde un diseñador de software escribe con el fin de dar un equipo de desarrollo de software orientado,
+                                generalmente, a la arquitectura de un proyecto.
                             </Typography>
                         </CardContent>
                     </Collapse>
                 </Card>
 
-
         </div>
-
-
-
-    );
+);
 }
